FEAT: CGIManager all
